Use yaml for agent config content
set_agents_config.py/get_agents_config.py currently read and write the
raw json that's inside the api body. This isn't very user friendly.

Change them to work directly with the agent config-as-yaml content
instead. get_agents_config.py will extract the yaml confg from the json
response and print it. set_agents_config.py will take the agent config
from the file on the command line, verify that it can be parsed as yaml,
wrap it in json, and PUT it.