Add missing dependency (#43)
* Add missing dependency

pyyaml now used for secure example files

* Add one-line install including virtualenv

* Update tests

The error for set_secure_system_falco_rules.py has changed.

Now that the system rules file is not per-customer, it can't be tied to
this test account. So just verify it can be fetched without inspecting
its contents. This also changes the result of creating default
policies/modifying policies a bit.