Add tests for new policy priorities apis
Use list_policies -o to get the current policies order, reverse the
order using "jq reverse", and set the order.

Use grep -b to find where in the policies list the policy "Write Below
Binary Dir" is, before and after setting the order. You should find that
the position of the policy moves to farther in the output e.g. the end.

Also use tee when fetching the falco rules file so if there's an error
we will see it.