Add policy priorities (#45)
* Add API wrappers to set/get policy priorities.

New methods get_policy_priorities/set_policy_priorites map to GET
/api/policies/priorities, PUT /api/policies/priorities, respsectively.

They work on raw json objects. Some minimal massaging of the objects
will be handled in the example programs.

* Update usage string.

It doesn't read the policy name from the cmdline, and reads the policy
description from stdin.

* Add priorities order to list_policies

Add the ability to return the order of policies only instead of the full
list of priorities. With -o/--order-only, only the list of ids is
returned.

Also ensure that the list of priorities is always sorted in priorities
order.

* Wrapper program to set policy order.

Example program to set the order of secure policies. Takes a simple list
of policy ids as returned by list_policies --order-only, and wraps it in
the necessary version/date object before using set_policy_priorities.

* Add tests for new policy priorities apis

Use list_policies -o to get the current policies order, reverse the
order using "jq reverse", and set the order.

Use grep -b to find where in the policies list the policy "Write Below
Binary Dir" is, before and after setting the order. You should find that
the position of the policy moves to farther in the output e.g. the end.

Also use tee when fetching the falco rules file so if there's an error
we will see it.

* Update monitor tests to use python-sdc-client acct

Move all of the monitor unit tests from .travis.yml into a script
test_monitor_apis.sh, and get rid of the (commented-out) tests for the
pip version of sdcclient.

In test_monitor_apis.sh, instead of relying on a hard-coded hostname,
start an agent on the fly using docker and use the current hostname for
get_data_advanced/create_sysdig_capture.

Also make changes to let multiple copies of test script run at once:

Many of the test programs relied on hard coded names of alerts,
dashboards, notification channels, etc. This caused problems when
running two copies of the test script at once, for example for the
travisci push and pr tests.

Fix this by making the various alert/dashboard/channel names
configurable in each example script, and using unique names in the
wrapper script that calls the individual scripts.