Created a script to enable/disable a list of alerts
Cloned the 'update_alerts.py' example to create a script that would take a list of alerts, and flip their enabled status. A customer had requested the ability to selectively enable / disable a list of alerts on a schedule, so I created this so they could run it on a cronjob. Will also add this as a FR.