Support falco rules variants (#73)
* Remove support for old falco integration

Remove support for endpoints that managed the pre-secure integration of
falco rules into the agent. This was removed from the agent in 0.70.0.

* Methods/examples to set/get falco rules files

Add the following methods to the python client:

 - {get,set}_default_falco_rules_files: a wrapper around the api endpoint
   /api/settings/defaultRulesFiles, handling PUT and GET.
 - load_default_falco_rules_files: load a collection of files on disk
   with a documented structure, returning a dict suitable for use in
   set_default_falco_rules_files()
 - save_default_falco_rules_files: given a dict from
   get_default_falco_rules_files, save it to a collection of files on disk
   with a documented structure

Also add example programs
{set,get}_secure_default_falco_rules.py. get_... has the ability to
either print the returned set of files directly or save them using a
--save <root dir path> option. Similarly, set_ has a --load <root dir
path> option to load files from disk to a dict for the PUT
/api/settings/defaultRulesFiles.

set_ also has some easier-to-use command line options that allow
setting a single file and tag.